+++ plugins/global-components.js
@@ -1,4 +1,5 @@
 import Vue from 'vue'
+import Autocomplete from '@/components/ui/Autocomplete/index.vue'
 import Box from '@/components/ui/Box/index.vue'
 import Button from '@/components/ui/Button/index.vue'
 import Input from '@/components/ui/Input/index.vue'
@@ -6,6 +7,7 @@ import Label from '@/components/ui/Label/index.vue'
 import Radio from '@/components/ui/Radio/index.vue'
 import Select from '@/components/ui/Select/index.vue'
 import Spinner from '@/components/ui/Spinner/index.vue'
+Vue.component('Autocomplete', Autocomplete)
 Vue.component('Box', Box)
 Vue.component('Button', Button)
 Vue.component('Input', Input)